update(multiLang): clan code
