+++ message/th.json
@@ -136,11 +136,11 @@
   },
   "MetaData": {
     "Project": {
-      "Title": "ไทย CodeWork Tech | Contact Us",
-      "Description": "ไทย Get in touch with us and make your dream come true. Let’s create something extraordinary together!",
+      "Title": "ไทย CodeWork Tech | Project",
+      "Description": "ไทย Explore our successful projects with valued clients, showcasing our expertise and dedication.",
       "OpenGraph": {
-        "Title": "CodeWork Tech | Contact Us",
-        "Description": "ไทย Get in touch with us and make your dream come true. Let’s create something extraordinary together!"
+        "Title": "CodeWork Tech | Project",
+        "Description": "ไทย Explore our successful projects with valued clients, showcasing our expertise and dedication."
       }
     },
     "Service": {
@@ -160,10 +160,10 @@
       }
     },
     "Blog": {
-      "Title": "ไทย CodeWork Tech | Blogs",
+      "Title": "ไทย CodeWork Tech | Blog",
       "Description": "ไทย Explore the latest insights, trends, and expert opinions in technology and business through our blog.",
       "OpenGraph": {
-        "Title": "ไทย CodeWork Tech | Blogs",
+        "Title": "ไทย CodeWork Tech | Blog",
         "Description": "ไทย Explore the latest insights, trends, and expert opinions in technology and business through our blog."
       }
     },
