--- src/features/Homepage/OurProjectSection/withOurPoject.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const withOurProject = (Component: React.FC) => {
-  const Hoc = () => {
-   
-    return <Component  />;
-  };
-
-  return Hoc;
-};
-export default withOurProject;
